Initial cut to move database structure
