+++ src/features/Backup_Spotify_Data.py
@@ -16,8 +16,10 @@ from datetime import datetime
 
 import src.General_Spotify_Helpers as gsh
 
-from src.helpers.decorators import *
-from src.helpers.Settings   import Settings
+from src.helpers.Database_Helpers   import DatabaseHelpers, DatabaseSchema
+from src.helpers.decorators         import *
+from src.helpers.Settings           import Settings
+
 
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: This function recursively traverses any list/ dict with any complexity to find and replace every 'None'
@@ -41,150 +43,30 @@ def replace_none(data, replace_with: str):
 
 
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
-DESCRIPTION: Finds the given types of an SQLite table columns and returns their associated python type to help us
-             verify data going into our DB.
-INPUT: db_conn - SQLite DB connection that we will grab our 'table' columns from.
-       table - SQLite table that we will grab column types of.
-OUTPUT: List of python types 
-"""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
-def get_column_types(db_conn, table: str) -> list:
-    with db_conn:
-        columns = db_conn.execute(f"PRAGMA table_info({table})").fetchall()
-
-    sqlite_type_mapping = {
-        'INTEGER': int,
-        'TEXT': str,
-        'REAL': float,
-        'BLOB': bytes,
-        'NUMERIC': float  # NUMERIC 'could' be an int, we don't use it anyways 
-    }
-
-    column_types = []
-    for col in columns:
-        col_name = col[1]   # Column name (not really needed for our case)
-        col_type = col[2].upper()  # Column type in SQLite schema
-        python_type = sqlite_type_mapping.get(col_type, str)  # Default to str if type is not found
-        column_types.append(python_type)
-
-    return column_types
-
-
-"""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: Class that handles creating a backup of the user's followed artists, playlists, and all their tracks.
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 class BackupSpotifyData(LogAllMethods):
     
-    def __init__(self, spotify, db_path: str=None, logger: logging.Logger=None) -> None:
+    def __init__(self, spotify, backup_db_path: str=None, logger: logging.Logger=None) -> None:
         self.spotify = spotify
         self.logger = logger if logger is not None else logging.getLogger()
-        self.db_path = db_path or f"{Settings.BACKUPS_LOCATION}{datetime.today().date()}.db"
+        self.vault_db = DatabaseHelpers(logger=self.logger)
         
-    @contextlib.contextmanager
-    def connect_db(self):
-        conn = sqlite3.connect(self.db_path)
-        try:
-            conn.execute("PRAGMA foreign_keys = ON;")
-            yield conn 
-            conn.commit()
-        finally:
-            conn.close()
-
-    """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""
-    DESCRIPTION: Inserts a variable amount of elements into a database table while verifying the types of your 'values'
-                 match that of the columns in your 'table'.
-    INPUT: table - What table we will insert into (str).
-           values - What data will be inserted into the table.
-           batch_size - How many 'values' we can add into a table at once for performance issues.
-    Output: N/A
-    """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''"""""" 
-    def _insert_many(self, table: str, values: Union[list[dict], list[tuple], list], batch_size: int=500) -> None:
-        if not values:
-            return  # No data to insert
-
-        with self.connect_db() as db_conn:
-            expected_types = get_column_types(db_conn, table)
+        snapshot_db_path = backup_db_path or f"{Settings.BACKUPS_LOCATION}{datetime.today().date()}.db"
+        self.snapshot_db = DatabaseHelpers(db_path = snapshot_db_path, schema=DatabaseSchema.SNAPSHOT
+                                           , logger=self.logger)
+       
+    def _clear_vault_playlists(self) -> None:
+        self.logger.info(f"\t Clearing Vault Playlists")
         
-        # Translates values into our format from either dict, list of tuples, or just list
-        data = [tuple(d.values()) for d in values] if type(values[0]) is dict \
-                else [(v,) for v in values] if type(values[0]) is not tuple else values 
-
-        for row in data:
-            for i, (val, expected_type) in enumerate(zip(row, expected_types)):
-                if not isinstance(val, expected_type):
-                    raise ValueError(f"'{val}' in column {i+1} of table '{table}' should be of type {expected_type}")
-
-        placeholders = ", ".join("?" for _ in data[0])
-        query = f"INSERT OR IGNORE INTO {table} VALUES ({placeholders})"
-
-        with self.connect_db() as db_conn:
-            for i in range(0, len(data), batch_size):
-                batch = data[i:i + batch_size]
-                db_conn.executemany(query, batch)
-
-    """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""
-    DESCRIPTION: Creates the necessary tables for our SQLite db. These include the artist table, playback table, tracks
-                 table, and the playlists-tracks many to many relationship table.
-    INPUT: N/A
-    Output: N/A
-    """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""
-    def _create_backup_data_db(self) -> None:
-        with self.connect_db() as db_conn:
-            db_conn.executescript("""
-                PRAGMA foreign_keys = ON;
-                
-                CREATE TABLE IF NOT EXISTS playlists (
-                    id text UNIQUE PRIMARY KEY,
-                    name text,
-                    description text
-                ) WITHOUT ROWID;
-
-                CREATE TABLE IF NOT EXISTS artists (
-                    id text UNIQUE PRIMARY KEY,
-                    name text
-                ) WITHOUT ROWID;
-
-                CREATE TABLE IF NOT EXISTS albums (
-                    id text UNIQUE PRIMARY KEY,
-                    name text,
-                    release_date text
-                ) WITHOUT ROWID;
-
-                CREATE TABLE IF NOT EXISTS tracks (
-                    id text UNIQUE PRIMARY KEY,
-                    name text,
-                    duration_ms integer,
-                    is_local integer,
-                    is_playable integer
-                ) WITHOUT ROWID;
-
-                CREATE TABLE IF NOT EXISTS followed_artists (
-                    id text PRIMARY KEY REFERENCES artists(id)
-                ) WITHOUT ROWID;
-
-                CREATE TABLE IF NOT EXISTS playlists_tracks (
-                    id_playlist text REFERENCES playlists(id),
-                    id_track text REFERENCES tracks(id)
-                );
-
-                CREATE TABLE IF NOT EXISTS tracks_artists (
-                    id_track text REFERENCES tracks(id),
-                    id_artist text REFERENCES artists(id),
-                    UNIQUE(id_track, id_artist)
-                );
-
-                CREATE TABLE IF NOT EXISTS tracks_albums (
-                    id_track text REFERENCES tracks(id),
-                    id_album text REFERENCES albums(id),
-                    UNIQUE(id_track, id_album)
-                );
-
-                CREATE TABLE IF NOT EXISTS albums_artists (
-                    id_album text REFERENCES albums(id),
-                    id_artist text REFERENCES artists(id),
-                    UNIQUE(id_album, id_artist)
-                );
-            """)
-
+        with self.vault_db.connect_db() as db_conn:
+            db_conn.execute("DELETE FROM playlists_tracks;")
+            db_conn.execute("DELETE FROM playlists;")
+    
+    def _insert_into_databases(self, table: str, values: Union[list[dict], list[tuple], list]) -> None:
+        self.vault_db.insert_many(table, values)
+        self.snapshot_db.insert_many(table, values)
+    
     """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""
     DESCRIPTION: Queries all followed artists by the user, inserts them into the database.
     INPUT: N/A
@@ -194,8 +76,8 @@ class BackupSpotifyData(LogAllMethods):
     def _add_followed_artists_to_db(self) -> None:
         artists = self.spotify.get_user_artists(info=["id", "name"])
         self.logger.info(f"\t Inserting {len(artists)} Artists")
-        self._insert_many("artists", artists)
-        self._insert_many("followed_artists", [artist['id'] for artist in artists])
+        self._insert_into_databases("artists", artists)
+        self._insert_into_databases("followed_artists", [artist['id'] for artist in artists])
         
     """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""
     DESCRIPTION: Fills all appropriate tables from the tracks from a given playlist.
@@ -233,13 +115,13 @@ class BackupSpotifyData(LogAllMethods):
             tracks_albums_entries.append((track['id'], track['album']['id']))
             album_artists_entries += [(track['album']['id'], artist['id']) for artist in track['album']['artists']]
             
-        self._insert_many("tracks", track_table_entries)
-        self._insert_many("albums", album_table_entries)
-        self._insert_many("artists", artist_table_entries)
-        self._insert_many("playlists_tracks", playlists_tracks_entries)
-        self._insert_many("tracks_artists", tracks_artists_entries)
-        self._insert_many("tracks_albums", tracks_albums_entries)
-        self._insert_many("albums_artists", album_artists_entries)
+        self._insert_into_databases("tracks", track_table_entries)
+        self._insert_into_databases("albums", album_table_entries)
+        self._insert_into_databases("artists", artist_table_entries)
+        self._insert_into_databases("playlists_tracks", playlists_tracks_entries)
+        self._insert_into_databases("tracks_artists", tracks_artists_entries)
+        self._insert_into_databases("tracks_albums", tracks_albums_entries)
+        self._insert_into_databases("albums_artists", album_artists_entries)
     
     """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""
     DESCRIPTION: Adds all user playlists into our database.
@@ -250,7 +132,7 @@ class BackupSpotifyData(LogAllMethods):
     def _add_user_playlists_to_db(self) -> None:
         user_playlists = self.spotify.get_user_playlists(info=["id", "name", "description"])
         self.logger.info(f"\t Inserting {len(user_playlists)} Playlists")
-        self._insert_many("playlists", user_playlists)
+        self._insert_into_databases("playlists", user_playlists)
         
         for playlist in user_playlists:
             self.logger.debug(f"\t Saving Data For Playlist: {playlist['name']}")
@@ -267,7 +149,6 @@ class BackupSpotifyData(LogAllMethods):
     """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""''""""""    
     def backup_data(self) -> None:
         self.logger.info(f"CREATING NEW BACKUP =====================================================================")
-        self._create_backup_data_db()
         self._add_followed_artists_to_db()
         self._add_user_playlists_to_db()
