+++ src/helpers/Settings.py
@@ -70,6 +70,7 @@ class SettingsClass:
 
     # DB Locations
     BACKUPS_LOCATION: str       = "databases/backups/"
+    LISTENING_VAULT_DB: str     = "databases/listening_vault.db"
     LAST_TRACK_PICKLE: str      = "databases/lastTrack.pk"
     LISTENING_DB: str           = "databases/listening_data.db"
     TRACK_COUNTS_DB: str        = "databases/track_counts.db"
